offline ir online/projektas -sql/requirements.txt
pygame
moviepy
mysql_connector